==> 0_datasets_merging.ipynb <==
import pandas as pd
import numpy as np

data_0 = pd.read_csv('./datasets/trumps-taiwan-call-tweets/dataset.csv', header=None)
data_1 = pd.read_csv('./datasets/lovesdata-trump-tweets-5-4-09-12-5-16/trumptweets1205-127.csv')
data_2 = pd.read_csv('./datasets/lovesdata-trump-tweets-5-4-09-12-5-16/2017_01_28_Trump_Tweets.csv')
data_3 = pd.read_csv('./datasets/lovesdata-trump-tweets-5-4-09-12-5-16/2016_12_05-TrumpTwitterAll.csv')

print(len(data_0))
print(len(data_1))
print(len(data_2))
print(len(data_3))

document_list = [data_1, data_2, data_3]
data_list = data_0.iloc[:, 0].tolist()

for i, j in enumerate(document_list):
    data_list += j.loc[:, 'Tweet'].tolist()

data_merged = pd.DataFrame({
    'Tweet':data_list
})

print(len(data_merged))

data_merged.to_csv('./datasets/tweet_text_merged_tab.csv', sep='\t', 
                   encoding='utf8', index=False)
