==> 2_remove_stop_words.ipynb <==
import nltk
import pandas as pd
import re, string

# f = pd.read_csv('./datasets/4_original_text_only.csv', sep='\t')
f = pd.read_csv('./datasets/6_retweet_text_only.csv', sep='\t')
print(len(f))
tweet = f.iloc[:, 0].tolist()
print(tweet[0])

# # 去除stop words, link, account
# nltk.download('stopwords')
# nltk_stopwords = nltk.corpus.stopwords.words('english')

result1 = []

# translation table
intab = "!@#$\"\':"
outtab = ""
for i in range(len(intab)):
    outtab += " "
trantab = str.maketrans(intab, outtab)


for i in range(0, len(tweet)):
    tweet_split_list = tweet[i].split(" ")
    words = []
    
    # convert all words inside the list to be lower case
    for j, k in enumerate(tweet_split_list):
        tweet_split_list[j] = k.lower()
    
    for j in range(0, len(tweet_split_list)):
        if j == 1:
            if '@' not in tweet_split_list[j]:
                if 'http' not in tweet_split_list[j]:
                    words.append(tweet_split_list[j].lower())
#                     if tweet_split_list[j] not in nltk_stopwords and tweet_split_list[j].lower() not in nltk_stopwords and tweet[i][j].upper() not in nltk_stopwords:
#                         words.append(tweet_split_list[j])
        else:
            if 'http' not in tweet_split_list[j]:
                tweet_split_list[j] =  tweet_split_list[j].translate(trantab).replace(" ", "")
                words.append(tweet_split_list[j])
#             if tweet[i][j] not in nltk_stopwords and tweet_split_list[j].lower() not in nltk_stopwords and tweet_split_list[j].upper() not in nltk_stopwords:
#                 if 'http' not in tweet_split_list[j]:
#                     words.append(tweet_split_list[j])
        
    result1.append(words)
    
print(len(result1))
print(result1[0])

# 去除標點符號
result2 = []
for i in range(0, len(result1)):
    words = []
    for j in range(0, len(result1[i])):
        out = ''.join(c for c in result1[i][j] if c not in string.punctuation)
        out = ''.join(x for x in out if ord(x) < 256)
        words.append(out)
    while '' in words:
        words.remove('')
    while 'rt' in words:
        words.remove('rt')
    while '\uea1d' in words:
        words.remove('\uea1d')
    result2.append(words)

print(len(result2))
print(result2[1:3])

# f = open('./datasets/7_retweet_text_only_stopwords_removed.csv', "w+", encoding="utf-8")
# for i in range(0, len(result2)):
#     print(result2[i])
#     s = ""
#     for j in range(0, len(result2[i])):
#         s += result2[i][j] + ' '
#     f.write(s)
#     f.write('\n')
# f.close()

content_list = []
for i, j in enumerate(result2):
    content = " ".join(str(b) for a, b in enumerate(j))
    content_list.append(content)

# read in the data with whole information
original_data = pd.read_csv("./datasets/5_retweet_text.csv", sep='\t')
# original_data = pd.read_csv("./datasets/3_original_text.csv", sep='\t')
print(len(original_data))
print(original_data.columns)

# add new column
original_data.loc[:, 'cleaned_retweet_text'] = content_list
# original_data.loc[:, 'cleaned_original_text'] = content_list
print(len(original_data))
print(original_data.head())

original_data.to_csv("./datasets/11_retweet_text_only_cleaned.csv", 
                     sep="\t",
                    index=None)
# original_data.to_csv("./datasets/12_original_text_only_cleaned.csv", 
#                      sep="\t",
#                     index=None)

import pandas as pd
import numpy as np

# data = pd.read_csv('./datasets/5_retweet_text.csv', sep='\t')
# print(data.columns)
data = pd.read_csv('./datasets/7_retweet_text_only_stopwords_removed.csv', sep='\t')
print(data.columns)

for i, j in enumerate(data.loc[:, 'stopwords_removed_retweet_text']):
    account = str(data.iloc[i, 1]).lower()
    str(j).replace(account, "")
    data.loc[i, 'without_account'] = j
    
print(data.head())
