==> 3_gensim_training.ipynb <==
# Data Preprocessing
* change to space splitted basis.
* remove duplicate rows and rows that consist too few element.

# Train Gensim Models - Word2Vec

from gensim import corpora
from gensim.models.word2vec import Word2Vec
from multiprocessing import cpu_count
from pprint import pprint
import gensim
import gensim.downloader as api
import numpy as np
import pandas as pd

# train word2vec model by genism, with data: text8
dataset = api.load('text8')
data = [d for d in dataset]

# train word2vec model. Default result vector size = 100
model = Word2Vec(data, min_count=0, workers=cpu_count())

model.save('./gensim_models/word2vec_model_text8')

# update our model with tweet dataset

# create copora dictionary
tweet_data = pd.read_csv('./datasets/tweet_stopwords_duplicates_short_removed_space_splitted.csv')
data_list = tweet_data.iloc[:, 0].tolist()
# tokenize data into words
texts = [[text for text in tweet.split()] for tweet in data_list]
# create dictionary by copora
dictionary = corpora.Dictionary(texts)
print(type(dictionary))
print(type(texts))


# load text8 word2vec dataset
model = Word2Vec.load('./gensim_models/word2vec_model_text8')
print(model['media'], '\n')

# update the model with tweets dataset
model.build_vocab(texts, update=True)
model.train(data, total_examples=model.corpus_count, epochs=model.iter)
print(model['media'])

# save the model
model.save('./gensim_models/word2vec_model_text8_tweet')

# train a word2vec model with tweet dataset only

# create copora dictionary
tweet_data = pd.read_csv('./datasets/tweet_stopwords_duplicates_short_removed_space_splitted.csv')
data_list = tweet_data.iloc[:, 0].tolist()
# tokenize data into words
texts = [[text for text in tweet.split()] for tweet in data_list]
# create dictionary by copora
dictionary = corpora.Dictionary(texts)
print(type(dictionary))
print(type(texts))

# train word2vec model. Default result vector size = 100
model = Word2Vec(texts, min_count=0, workers=cpu_count())
print(model['trump'], '\n')
print(model['taiwan'])

model.save('./gensim_models/word2vec_model_tweet')

# check the result with only tweets data
print(model.wv.most_similar('trump'), '\n')
print(model.wv.most_similar('taiwan'), '\n')
print(model.wv.most_similar('obama'), '\n')
print(model.wv.most_similar('freak'), '\n')

# Gensim Training - Doc2Vec

import gensim
import gensim.downloader as api
import numpy as np
import pandas as pd

# function to create 'tagged document'
def create_tagged_document(list_of_list_of_word):
    for i, list_of_word in enumerate(list_of_list_of_word):
        yield gensim.models.doc2vec.TaggedDocument(list_of_word, [i])

# create 'tagged document' dataset of text8
# load text8 dataset
dataset = api.load('text8')
data = [d for d in dataset]

# create tagged document of text8
training_data_text8 = list(create_tagged_document(data))

# initialize doc2vec model
model = gensim.models.doc2vec.Doc2Vec(vector_size=50, min_count=2, epochs=100)

# build the vocabulary
model.build_vocab(training_data_text8)

# train the Doc2Vec model
model.train(training_data_text8, total_examples=model.corpus_count, epochs=model.epochs)

model.save('./gensim_models/doc2vec_model_text8')

# create a dataset with text8 and tweets data
tweet_data = pd.read_csv('./datasets/tweet_stopwords_duplicates_short_removed_space_splitted.csv')
data_list = tweet_data.iloc[:, 0].tolist()

# tokenize data into words
training_tweet = [[text for text in tweet.split()] for tweet in data_list]

# create tagged document with text8 and tweet
training_data_text8_tweet = list(create_tagged_document(data + training_tweet))

# create tagged document with tweet only
training_data_tweet = list(create_tagged_document(training_tweet))

# create a model for both text8 and tweets data
model = gensim.models.doc2vec.Doc2Vec(vector_size=50, min_count=2, epochs=100)

# build the vocabulary
model.build_vocab(training_data_text8_tweet)

# train the model
model.train(training_data_text8_tweet, total_examples=model.corpus_count, epochs=model.epochs)

model.save('./gensim_models/doc2vec_model_text8_tweet')

# Create Tagged dataset and gensim models

import gensim
import gensim.downloader as api
import numpy as np
import pandas as pd

# function to create 'tagged document'
def create_tagged_document(list_of_list_of_word):
    for i, list_of_word in enumerate(list_of_list_of_word):
        yield gensim.models.doc2vec.TaggedDocument(list_of_word, [i])

# prepare data
original = pd.read_csv("./datasets/12_original_text_only_cleaned.csv", sep='\t')
# retweet = pd.read_csv("./datasets/11_retweet_text_only_cleaned.csv", sep='\t')
retweet = pd.read_csv("./datasets/20_retweet_data.csv", sep=",", encoding='ANSI')

print(len(original))
print(original.columns)
print(len(retweet))
print(retweet.columns)

# create amodel with both text8 and original tweet data

# create 'tagged document' dataset of text8
# load text8 dataset
dataset = api.load('text8')
data = [d for d in dataset]

# tokenize data into words
training_tweet = [[text.lower() for text in str(tweet).split()] for tweet in retweet.loc[:, 'retweet_text']]
# training_tweet = [[text for text in str(tweet).split()] for tweet in original.loc[:, 'cleaned_original_text']]

# create tagged document with tweet only
training_data_tweet = list(create_tagged_document(training_tweet + data))

# create a model for only tweets
model = gensim.models.doc2vec.Doc2Vec(vector_size=50, min_count=2, epochs=100)

# build the vocabulary
model.build_vocab(training_data_tweet)

# train the model
model.train(training_data_tweet, total_examples=model.corpus_count, epochs=model.epochs)

model.save('./gensim_models/doc2vec_model_retweet_new_text8')
# model.save('./gensim_models/doc2vec_model_original_new')

# tokenize data into words
training_tweet = [[text for text in str(tweet).split()] for tweet in retweet.loc[:, 'cleaned_retweet_text']]
# training_tweet = [[text for text in str(tweet).split()] for tweet in original.loc[:, 'cleaned_original_text']]

# create tagged document with tweet only
training_data_tweet = list(create_tagged_document(training_tweet))

# create a model for only tweets
model = gensim.models.doc2vec.Doc2Vec(vector_size=50, min_count=2, epochs=100)

# build the vocabulary
model.build_vocab(training_data_tweet)

# train the model
model.train(training_data_tweet, total_examples=model.corpus_count, epochs=model.epochs)

model.save('./gensim_models/doc2vec_model_retweet_new')
# model.save('./gensim_models/doc2vec_model_original_new')

# Create dataset with Doc2Vec Features

import gensim
import numpy as np
import pandas as pd
from tqdm import tqdm

# model = gensim.models.doc2vec.Doc2Vec.load('./gensim_models/doc2vec_model_retweet_new')
model = gensim.models.doc2vec.Doc2Vec.load('./gensim_models/doc2vec_model_retweet_new_text8')

for i, j in enumerate(tqdm(retweet.loc[:, 'retweet_text'])):
    element_splitted = str(j).lower().split(" ")
    feature_array = model.infer_vector(element_splitted)
    for a, b in enumerate(feature_array):
        column_name = 'feature_' + str(a)
        retweet.loc[i, column_name] = b

print(retweet.head())
print(retweet.describe())
print(retweet.shape)

print(retweet.loc[541, 'feature_0':])

# retweet.to_csv('./datasets/13_retweet_text_doc2vec_new.csv', sep='\t', header=True, index=None)
retweet.to_csv('./datasets/19_retweet_text_doc2vec_new_text8.csv', sep='=', header=True, index=True)

model = gensim.models.doc2vec.Doc2Vec.load('./gensim_models/doc2vec_model_original_new')

for i, j in enumerate(tqdm(original.loc[:, 'cleaned_original_text'])):
    element_splitted = str(j).split(" ")
    feature_array = model.infer_vector(element_splitted)
    for a, b in enumerate(feature_array):
        column_name = 'feature_' + str(a)
        original.loc[i, column_name] = b

print(original.head())
print(original.describe())
original.to_csv('./datasets/14_original_text_doc2vec_new.csv', sep='\t', header=True, index=None)

# Gensim Vector Relations Check

import gensim
import numpy as np
import pandas as pd
from tqdm import tqdm

# checking words similarities

model = gensim.models.doc2vec.Doc2Vec.load('./gensim_models/doc2vec_model_retweet_new')

print('happy\n', model.wv.most_similar('happy'), '\n')
print('nytimes\n', model.wv.most_similar('nytimes'), '\n')
print('trump\n', model.wv.most_similar('trump'), '\n')
print('realdonaldtrump\n', model.wv.most_similar('trump'), '\n')
print('obama\n', model.wv.most_similar('obama'), '\n')
print('taiwan\n', model.wv.most_similar('taiwan'), '\n')
print('china\n', model.wv.most_similar('china'), '\n')
print('beijing\n', model.wv.most_similar('beijing'), '\n')
print('cnn\n', model.wv.most_similar('cnn'), '\n')

# print out vector of words in our dataset

print(model['trump'], '\n')
print(model['taiwan'], '\n')

# reference: https://stackoverflow.com/questions/42781292/doc2vec-get-most-similar-documents
word_tokens = "the nytimes was thrilled by obama talking to communist dictator castro and horrified by trump talking to the elected le".split(" ")
token_vector = model.infer_vector(word_tokens)
sims = model.docvecs.most_similar([token_vector])
print(sims)

# print out the original text
data = pd.read_csv("./datasets/13_retweet_text_doc2vec_new.csv", sep='\t')
print("the nytimes was thrilled by obama talking to communist dictator castro and horrified by trump talking to the elected le")
print('\n', data.loc[183, 'retweet_text'], '\n')
print(data.loc[584, 'retweet_text'], '\n')
print(data.loc[118, 'retweet_text'], '\n')
print(data.loc[455, 'retweet_text'], '\n\n')


word_tokens = "incompetence is really the primary threat from the trump administration".split(" ")
token_vector = model.infer_vector(word_tokens)
sims = model.docvecs.most_similar([token_vector])
print(sims)

# print out the original text
data = pd.read_csv("./datasets/13_retweet_text_doc2vec_new.csv", sep='\t')
print('\n', data.loc[38, 'retweet_text'], '\n')
print(data.loc[569, 'retweet_text'], '\n')
print(data.loc[40, 'retweet_text'], '\n')
print(data.loc[324, 'retweet_text'], '\n\n')


word_tokens = "obama turned his back on taiwan".split(" ")
token_vector = model.infer_vector(word_tokens)
sims = model.docvecs.most_similar([token_vector])
print(sims)

# print out the original text
data = pd.read_csv("./datasets/13_retweet_text_doc2vec_new.csv", sep='\t')
print('\n', data.loc[128, 'retweet_text'], '\n')
print(data.loc[579, 'retweet_text'], '\n')
print(data.loc[134, 'retweet_text'], '\n')
print(data.loc[162, 'retweet_text'], '\n\n')

data = pd.read_csv('./datasets/10_original_text_doc2vec.csv', sep='\t')

# cosine similarity search

from sklearn.datasets import fetch_20newsgroups
twenty = fetch_20newsgroups()
print(type(twenty.data))
print(twenty.data[0])
# tfidf = TfidfVectorizer().fit_transform(twenty.data)
