==> 6_Creating_tfidf_dataset.ipynb <==
# Create tfidf Vector Dataset

from sklearn.feature_extraction.text import TfidfVectorizer
import numpy as np
import pandas as pd
import string

# data = pd.read_csv("./datasets/7_retweet_text_only_stopwords_removed.csv", 
#                   sep='\t')
data = pd.read_csv("./datasets/8_original_text_only_stopwords_removed.csv", 
                  sep='\t')

# translation table
intab = "!@#$\"\':\r\n"
outtab = ""
for i in range(len(intab)):
    outtab += "@"
trantab = str.maketrans(intab, outtab)

# print(data.head())
print(data.columns)
# text_list = data.loc[:,  'stopwords_removed_retweet_text'].tolist()
text_list = data.loc[:,  'stopwords_removed_original_text'].tolist()

for i, j in enumerate(text_list):
    text_list[i] =  str(j).translate(trantab).replace("@", "")
    
print(text_list[:10])

# create tfidf vector
vectorizer = TfidfVectorizer(stop_words='english')

tfidf_vector = vectorizer.fit_transform(text_list)

print(tfidf_vector.shape)

tfidf_vectorize_data = pd.DataFrame(
    data = tfidf_vector.toarray(),
    columns = ['feature_' + str(i) for i in range(tfidf_vector.shape[1])]
)

print(tfidf_vectorize_data.shape)

# concat original dataset
data = pd.concat([data, tfidf_vectorize_data], axis=1)

print(data.columns)
print(data.shape)

# data.to_csv("./datasets/15_retweet_tfidf_vectorized.csv", sep='\t', encoding='utf8', index=True)
data.to_csv("./datasets/16_original_tfidf_vectorized.csv", sep='\t', encoding='utf8', index=True)
